removing business component for deleteIdentity
